/retweet revised

+++ twitter-server/src/main/java/com/example/twitter/service/RetweetServiceImpl.java
@@ -1,67 +1,53 @@
 package com.example.twitter.service;
 
-import com.example.twitter.entity.Comment;
+import com.example.twitter.dto.RetweetRequestDto;
+import com.example.twitter.dto.RetweetResponseDto;
 import com.example.twitter.entity.Retweet;
 import com.example.twitter.entity.Tweet;
 import com.example.twitter.entity.User;
+import com.example.twitter.exceptions.RetweetExistsException;
 import com.example.twitter.exceptions.RetweetNotFoundException;
+import com.example.twitter.exceptions.UserNotFoundException;
 import com.example.twitter.repository.RetweetRepository;
+import com.example.twitter.repository.UserRepository;
+import com.example.twitter.utils.DtoMapping;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class RetweetServiceImpl implements RetweetService{
 
+    private DtoMapping dtoMapping;
+    private UserRepository userRepository;
     private RetweetRepository retweetRepository;
 
     @Autowired
-    public RetweetServiceImpl(RetweetRepository retweetRepository){
+    public RetweetServiceImpl(DtoMapping dtoMapping, UserRepository userRepository, RetweetRepository retweetRepository){
+        this.dtoMapping = dtoMapping;
+        this.userRepository = userRepository;
         this.retweetRepository = retweetRepository;
     }
 
     @Override
-    public List<Retweet> getALl() {
-        return retweetRepository.findAll();
-    }
-
-    @Override
-    public Retweet getById(Long id) {
-        return retweetRepository
-                .findById(id)
-                .orElseThrow(()->new RetweetNotFoundException("Retweet with " + id + " not found."));
-    }
-
-    @Override
-    public Retweet save(Retweet retweet, Comment comment, Tweet tweet, User user) {
-        user.addRetweet(retweet);
-        if(tweet != null) tweet.addRetweet(retweet);
-        if(comment != null) comment.addRetweet(retweet);
-        return retweetRepository.save(retweet);
-    }
-
-    @Override
-    public Retweet put(Long id, Retweet retweet, Comment comment, Tweet tweet, User user) {
-        Optional<Retweet> retweetOptional = retweetRepository.findById(id);
-        if(retweetOptional.isPresent()){
-            retweet.setId(id);
-            return retweetRepository.save(retweet);
+    public RetweetResponseDto save(RetweetRequestDto retweetRequestDto, String username) {
+        if(retweetRepository.getRetweetOfTweetByTweetIdAndUsername(retweetRequestDto.getTweetId(), username) != null){
+            throw new RetweetExistsException("Already retweeted.");
         }
-        user.addRetweet(retweet);
-        if(tweet != null) tweet.addRetweet(retweet);
-        if(comment != null) comment.addRetweet(retweet);
-        return retweetRepository.save(retweet);
-    }
-
-    @Override
-    public Retweet patch(Long id, Retweet retweet) {
-        return retweet;
+        User user = userRepository
+                .findUserByUsername(username)
+                .orElseThrow(()-> new UserNotFoundException("User with username: " + username + " not found."));
+        Retweet retweet = dtoMapping.MappingRetweetRequestToRetweet(retweetRequestDto);
+        retweet.setUser(user);
+        retweetRepository.save(retweet);
+        return dtoMapping.MappingRetweetToRetweetResponseDto(retweet);
     }
 
     @Override
     public void delete(Long id) {
+        if(retweetRepository.findById(id).isEmpty()){
+            throw new RetweetNotFoundException("Retweet with id: " + id + " not found.");
+        }
         retweetRepository.deleteById(id);
     }
 }
